pyproject.toml
[tool.poetry]
name = "jackal"
version = "0.1.0"
description = ""
authors = ["Koki Kubo <c11h17n6o8@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
siegeapi = "^3.0.2"
matplotlib = "^3.5.3"
pandas = "^1.4.3"
python-dotenv = "^0.20.0"
typer = "^0.6.1"
"discord.py" = {git = "https://github.com/Rapptz/discord.py", rev = "master"}
pendulum = "^2.1.2"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
